1_sqlite_mcp_agent/requirements.txt
fastmcp==2.13.2
llama-index==0.14.9
llama-index-core==0.14.9
llama-index-llms-gemini==0.6.1
llama-index-tools-mcp==0.4.3
python-dotenv==1.2.1
google-generativeai==0.8.5
